Split monolithic Zustand store into focused auth/chat/ui stores
Performance improvement: the ~1700-line single store caused every
subscribed component to check for re-renders on any state change.
Now state is split into three domain-focused stores:

- authStore: user, session, orgs, masquerade (persisted)
- chatStore: conversations, messages, streaming, integrations
- uiStore: sidebar, current view, pinned chats (persisted)

A backward-compatible facade (useAppStore) merges all three so
existing consumers continue to work unchanged. Components can
opt into direct sub-store imports for optimal re-render behavior.

Includes localStorage migration from the legacy "revtops-store"
key to the new split store keys.
