+++ frontend/src/store/types.ts
@@ -1,0 +1,257 @@
+/**
+ * Shared type definitions for the store layer.
+ *
+ * Extracted into a separate module so that sub-stores can import types
+ * without circular dependencies.
+ */
+
+// =============================================================================
+// Auth types
+// =============================================================================
+
+export interface UserProfile {
+  id: string;
+  email: string;
+  name: string | null;
+  avatarUrl: string | null;
+  agentGlobalCommands: string | null;
+  phoneNumber: string | null;
+  jobTitle: string | null;
+  roles: string[]; // Global roles like ['global_admin']
+}
+
+export interface MasqueradeState {
+  originalUser: UserProfile;
+  originalOrganization: OrganizationInfo | null;
+  masqueradingAs: UserProfile;
+  masqueradeOrganization: OrganizationInfo | null;
+}
+
+export interface OrganizationInfo {
+  id: string;
+  name: string;
+  logoUrl: string | null;
+  handle?: string | null; // Optional for backwards compat with persisted state
+}
+
+export interface UserOrganization {
+  id: string;
+  name: string;
+  logoUrl: string | null;
+  handle?: string | null;
+  role: string;
+  isActive: boolean;
+}
+
+// =============================================================================
+// Integration types (data sources)
+// =============================================================================
+
+export interface TeamConnection {
+  userId: string;
+  userName: string;
+}
+
+export interface SyncStats {
+  accounts?: number;
+  deals?: number;
+  contacts?: number;
+  activities?: number;
+  channels?: number;
+  pipelines?: number;
+  goals?: number;
+  repositories?: number;
+  commits?: number;
+  pull_requests?: number;
+  total_files?: number;
+  docs?: number;
+  sheets?: number;
+  slides?: number;
+  folders?: number;
+  // Issue tracker providers (Linear, Jira, Asana)
+  teams?: number;
+  projects?: number;
+  issues?: number;
+}
+
+export interface Integration {
+  id: string;
+  provider: string;
+  userId: string | null;
+  isActive: boolean;
+  lastSyncAt: string | null;
+  lastError: string | null;
+  connectedAt: string | null;
+  connectedBy: string | null;
+  scope: "organization" | "user";
+  // Sharing settings
+  shareSyncedData: boolean;
+  shareQueryAccess: boolean;
+  shareWriteAccess: boolean;
+  pendingSharingConfig: boolean;
+  // Ownership
+  isOwner: boolean;
+  currentUserConnected: boolean;
+  teamConnections: TeamConnection[];
+  teamTotal: number;
+  syncStats: SyncStats | null;
+}
+
+// =============================================================================
+// Chat / conversation types
+// =============================================================================
+
+export interface Participant {
+  id: string;
+  name: string | null;
+  email: string;
+  avatarUrl?: string | null;
+}
+
+export interface ChatSummary {
+  id: string;
+  title: string;
+  lastMessageAt: Date;
+  previewText: string;
+  type?: "agent" | "workflow";
+  workflowId?: string;
+  scope: "private" | "shared";
+  userId?: string; // Creator's user ID
+  participants?: Participant[];
+}
+
+// Content block types (matches API)
+export interface TextBlock {
+  type: "text";
+  text: string;
+}
+
+export interface ToolUseBlock {
+  type: "tool_use";
+  id: string;
+  name: string;
+  input: Record<string, unknown>;
+  result?: Record<string, unknown>;
+  status?: "pending" | "running" | "complete" | "streaming";
+}
+
+export interface ErrorBlock {
+  type: "error";
+  message: string;
+}
+
+export interface ArtifactBlock {
+  type: "artifact";
+  artifact: {
+    id: string;
+    title: string;
+    filename: string;
+    contentType: "text" | "markdown" | "pdf" | "chart";
+    mimeType: string;
+  };
+}
+
+export interface AppBlock {
+  type: "app";
+  app: {
+    id: string;
+    title: string;
+    description: string | null;
+    frontendCode: string;
+    frontendCodeCompiled?: string | null;
+  };
+}
+
+export interface AttachmentBlock {
+  type: "attachment";
+  filename: string;
+  mimeType: string;
+  size: number;
+}
+
+export type ContentBlock =
+  | TextBlock
+  | ToolUseBlock
+  | ErrorBlock
+  | ArtifactBlock
+  | AppBlock
+  | AttachmentBlock;
+
+// Legacy type for streaming compatibility
+export interface ToolCallData {
+  toolName: string;
+  toolId: string;
+  input: Record<string, unknown>;
+  result?: Record<string, unknown>;
+  status: "running" | "complete" | "error";
+}
+
+export interface ChatMessage {
+  id: string;
+  role: "user" | "assistant";
+  contentBlocks: ContentBlock[];
+  timestamp: Date;
+  isStreaming?: boolean;
+  userId?: string;
+  senderName?: string | null;
+  senderEmail?: string | null;
+  senderAvatarUrl?: string | null;
+}
+
+// Pending chunk for out-of-order handling
+export interface PendingChunk {
+  index: number;
+  content: string;
+}
+
+// Per-conversation summary (generated by backend)
+export interface ConversationSummaryData {
+  overall: string;
+  recent: string;
+  message_count_at_generation: number;
+  updated_at: string;
+}
+
+// Per-conversation state
+export interface ConversationState {
+  messages: ChatMessage[];
+  title: string;
+  isThinking: boolean;
+  streamingMessageId: string | null;
+  activeTaskId: string | null;
+  lastChunkIndex: number;
+  pendingChunks: PendingChunk[];
+  summary: ConversationSummaryData | null;
+  hasMore: boolean;
+  contextTokens: number | null;
+}
+
+// Task state from backend
+export interface ActiveTask {
+  id: string;
+  conversation_id: string;
+  status: string;
+  output_chunks: Array<{
+    index: number;
+    type: string;
+    data: unknown;
+    timestamp: string;
+  }>;
+}
+
+// =============================================================================
+// UI types
+// =============================================================================
+
+export type View =
+  | "home"
+  | "chat"
+  | "data-sources"
+  | "data"
+  | "workflows"
+  | "memory"
+  | "apps"
+  | "app-view"
+  | "artifact-view"
+  | "admin"
+  | "pending-changes";
